+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # RELEASED
 
+## 0.2.0 (Aug 6th, 2025)
+
+- Remove `BufMut` implementation for `BytesMut` and `Vec<u8>`
+- Add `advance_mut` and `try_advance_mut` for `BufMut`
+- Add `write_*` and `try_write_*` APIs
+
 ## 0.1.4 (Aug 5th, 2025)
 
 - Add `buffer_from` and `buffer_from_checked`